changed the logging

+++ check.py
@@ -7,6 +7,8 @@ import sys
 from subprocess import call
 from subprocess import check_call
 
+logText = ""
+success = True
 
 call(["getmail",])
 
@@ -19,42 +21,79 @@ for mail in newmails:
 		with open(mail, 'r') as f:
   			first_line = f.readline()
   			sender = re.search("<([\w\-][\w\-\.]+@[\w\-][\w\-\.]+[a-zA-Z]{1,4})>", first_line).group(1)
+				logText += "Successful parse:\n"
+				logText += sender + "/n"
+	except:
+		logText += "Error in parsing %s\n" % mail
+		success = False
+		break
+
+	try:
 		# get image
 		cmd = "munpack -C /home/pi/nodescan " + mail
 		os.system(cmd)
-		images = glob.glob("/home/pi/nodescan/*.jpg") + glob.glob("/home/pi/nodescan/*.JPG")
-		# if image then process
-		if len(images) > 0:
+		logText += "Succesfull outputting image munpack \n"
+	except:
+		logText += "Error in outputting image munpack \n"
+		success = False
+		break
+
+
+	images = glob.glob("/home/pi/nodescan/*.jpg") + glob.glob("/home/pi/nodescan/*.JPG")
+	# if image then process
+	if len(images) > 0:
+		try:
 			check_call(["python", scandir + "scan.py", "--image", images[0]])
+			logText += "Successfully used scan.py \n"
+		except:
+			logText += "Problem with scan.py \n"
+			success = False
+			break
+
+		try:
 			# send back the ok message with the file attached
 			cmd = "mutt -s 'Voila ton scan !' -i /home/pi/nodescan/ok_msg.txt " + sender +  " -a /home/pi/*.pdf < /dev/null"
 			os.system(cmd)
-						
-		else:
-			# there is no image to process
-			# send back the "no image attached" error message
-			cmd = "mutt -s 'Euh, elle est ou ton image ?' -i /home/pi/nodescan/error_empty_msg.txt " + sender +  " < /dev/null"
-			os.system(cmd)
-		
+			logText += "Successfully send the scaned image \n")
+		except:
+			logText += "Problem sending mail \n")
+			success = False
+			break
+	else:
+		# there is no image to process
+		# send back the "no image attached" error message
+		cmd = "mutt -s 'Euh, elle est ou ton image ?' -i /home/pi/nodescan/error_empty_msg.txt " + sender +  " < /dev/null"
+		os.system(cmd)
+		logText += "no image in mail \n")
+		success = False
+		break
 
+	try:
 		# clean
 		os.remove(mail)
 		os.system("rm -f /home/pi/nodescan/*.JPG /home/pi/nodescan/*.jpg /home/pi/nodescan/*.desc /home/pi/*.pdf")
 		os.system("rm -f /home/pi/scanMail/sent/cur/*")
-		#call(["rm part*"])
-		#call(["rm "+mail,])
-
+		logText += "Successfully cleaned files \n"
 	except:
-		exc_type, exc_obj, exc_tb = sys.exc_info()
-		fname = os.path.split(exc_tb.tb_frame.f_code.co_filename)[1]
-		print(exc_type, fname, exc_tb.tb_lineno)
-		print "pb with %s" % mail
+		logText += "Problem in cleaning files \n"
+		success = False
+		break
 
+
+	if success != False:
 		# the image was not correctly processed
 		# send back the "file not processed" error message
-		cmd = "mutt -s 'Elle est moche ta photo' -i /home/pi/nodescan/error_process_msg.txt " + sender +  " < /dev/null"
+		tmp = open("tmp.txt", "w")
+		tmp.write(logText)
+		tmp.close()
+		cmd = "mutt -s 'Problem de scan...' -i /home/pi/nodescan/tmp.txt " + sender +  " < /dev/null"
 		os.system(cmd)
 		# clean
 		os.remove(mail)
 		os.system("rm -f /home/pi/nodescan/*.JPG /home/pi/nodescan/*.jpg /home/pi/nodescan/*.desc /home/pi/*.pdf")
 		os.system("rm -f /home/pi/scanMail/sent/cur/*")
+
+
+logs = open("log.txt", "a")
+logs.write(logText)
+logs.close()
